Se agregaron los logs y se esta probando el envio de mails

+++ backEnd/cinema/cinema/src/main/java/com/backend/cinema/controllers/MovieController.java
@@ -1,6 +1,9 @@
 package com.backend.cinema.controllers;
+
 import java.util.Set;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,51 +13,70 @@ import com.backend.cinema.dto.MovieDTO;
 import com.backend.cinema.exception.ResourceNotFoundException;
 import com.backend.cinema.services.impl.MovieServiceImpl;
 
-
 @RestController
 @RequestMapping("/movie")
-public class MovieController{
-    //public static final Logger log = LogManager.getLogger(MovieServiceImpl.class);
+public class MovieController {
+    public static final Logger log = LogManager.getLogger(MovieServiceImpl.class);
     private MovieServiceImpl movieService;
 
     @Autowired
-    public MovieController(MovieServiceImpl movieService){
+    public MovieController(MovieServiceImpl movieService) {
         this.movieService = movieService;
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getId(@PathVariable Long id) throws ResourceNotFoundException{
+    public ResponseEntity<?> getId(@PathVariable Long id) throws ResourceNotFoundException {
         MovieDTO movieDTO = movieService.getId(id);
+        if (movieDTO == null) {
+            log.error("Movie not found with ID: {}", id);
+            throw new ResourceNotFoundException("Error retrieving movie.");
+        }
+        log.info("Movie successfully retrieved with ID: {}", id);
         return ResponseEntity.ok().body(movieDTO);
     }
 
     @GetMapping()
-    public ResponseEntity<Set<MovieDTO>> getMovies() throws ResourceNotFoundException{
-        return new ResponseEntity<>(movieService.getAll(), HttpStatus.OK);
+    public ResponseEntity<Set<MovieDTO>> getMovies() throws ResourceNotFoundException {
+        Set<MovieDTO> movies = movieService.getAll();
+        if (movies.isEmpty()) {
+            log.warn("No movies found");
+            throw new ResourceNotFoundException("No movie found");
+        }
+        log.info("Movies found: {}", movies);
+        return new ResponseEntity<>(movies, HttpStatus.OK);
     }
 
     @PostMapping()
-    public ResponseEntity<MovieDTO> save(@RequestBody MovieDTO movieDTO){
+    public ResponseEntity<MovieDTO> save(@RequestBody MovieDTO movieDTO) {
+        log.info("Creating user:", movieDTO);
         MovieDTO response = movieService.save(movieDTO);
         return ResponseEntity.ok().body(response);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable Long id){
-        ResponseEntity response = null;
-        if (id != 0) response = ResponseEntity.status(HttpStatus.OK).build();
-        else response = ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        movieService.delete(id);
-        return response;
+    public ResponseEntity<?> delete(@PathVariable Long id) {
+        try {
+            if (!movieService.existsById(id)) {
+                log.warn("Movie not found with ID: {}", id);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+            movieService.delete(id);
+            log.info("Movie deleted with ID: {}", id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+       } catch (IllegalArgumentException e) {
+            log.warn("Invalid ID: {}", id);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
     }
 
     @PutMapping
-    public void update(@RequestBody MovieDTO movieDTO){
-        movieService.update(movieDTO); 
+    public ResponseEntity<MovieDTO> update(@PathVariable Long id, @RequestBody MovieDTO movieDTO){
+        MovieDTO movie = movieService.save(movieDTO);
+        return new ResponseEntity<>(movie, HttpStatus.OK);
     }
 
     @ExceptionHandler(ResourceNotFoundException.class)
-    public String ProcessResourceNotFoundException(ResourceNotFoundException resourceNotFoundException){
+    public String ProcessResourceNotFoundException(ResourceNotFoundException resourceNotFoundException) {
         return resourceNotFoundException.getMessage();
-    }  
+    }
 }
